backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
motor==3.3.2
pydantic==1.10.26
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.0.1
python-multipart==0.0.6